V0.9 - Added items spawn manager and fixed bugs with platforms collision. Need to fix bug with platform top collision and finish item spawn manager.

+++ SpaceRunner/src/level/ChunkPlatform.java
@@ -6,7 +6,7 @@ public class ChunkPlatform extends Platform{
 	
 	private ChunkPlatform next_plt = null;
 	
-	public ChunkPlatform(float x, float y, int size, BufferedImage[] level_plts) {
+	public ChunkPlatform(int x, int y, int size, BufferedImage[] level_plts) {
 		super(x, y, size, level_plts);
 		
 		platformInit();
